+++ src/parser.hpp
@@ -3,9 +3,9 @@
 #include "token.hpp"
 #include <vector>
 
-int factor(const std::vector<Token>& tokens);
-int term(const std::vector<Token>& tokens);
-int expression(const std::vector<Token>& tokens);
+double factor(const std::vector<Token>& tokens);
+double term(const std::vector<Token>& tokens);
+double expression(const std::vector<Token>& tokens);
 
 void parse(const std::vector<Token>& tokens);
